Merging the soot-2-jsrfix branch into trunk.
This provides an interim fix to jsr-inlining in
soot/coffi/CFG.java, so that exception ranges are computed
correctly for inlined subroutines which partially overlap a
protected area.

This is an interim fix only, because it doesn't properly handle
protected areas which extend to the very end of the method
body.  Until we come up with a convention for what
Trap.getEndUnit() should return for traps that extend to the
end of the method, we simply lie, and cut the protected area
one statement short (thanks to Ondrej for that suggestion).