When inserting a Trap whose protected area extends up to an exception handler, adjust the area's endpoint to omit the @caughtException identity statement inserted at the beginning of that handler.
This bug came to our attention because two of the diffs included
in revision 1126 made the @caughtException identity appear within
the protected area of some exceptions which would not have
included it before 1126, but the pre-1126 code would also include
the @caughtException in other, less common situations (where the
handler at the end of the protected area is associated with a
different trap than the protected area).  So rather than
reverting the two troublesome diffs, this revision introduces a
different method for adjusting the protected area around the
@caughtException identities.

I suspect there may be other circumstances where this fix might
also fail, but have yet to come up with a failing test case.