Fix a bug which treats nested exception handlers which catch only some of the elements of AnySubType(t) as if they caught all of its elements.
This fix is about to be replaced by a reimplementation of parts
of soot.toolkits.exceptions.ThrowableSet and
soot.toolkits.exceptions.ExceptionalUnitGraph (in fact, that
reimplementation revealed the bug). I'm committing this change
first so that if the reimplementation has its own bugs,
withdrawing it will not reintroduce the nested exception handler
bug.