fix for bug 99:
"In older versions (probably more than a year ago), when exactly one class was
specified on the Soot command line, Soot would use that class as the main class
of the program. If 0 or multiple classes were specified on the command line,
Soot would require that a main class be explicitly specified using the
-main-class option.

The current Soot trunk chooses a seemingly random class as the main class. Even
when only a single class is given on the command line, Soot sometimes chooses a
different class as the main class. The chosen class is even non-deterministic
from one run of Soot to another (with identical command lines).

Could the original behaviour be restored?"


Fixed by first iterating through argument classes, then through all other classes.