Committing fir for infinite recursion in presence of phantom references. (see post "I got a heap memory error")
As Ben Bellamy writes:

The infinite loop was being caused by the piece of code that determines the least common ancestors of an array type and a reference type, in particular lines 148-151 of BytecodeHierarchy.java. All array types descend from Serializable and Cloneable, therefore if the reference type descends from either Serializable or Cloneable then it is a least common ancestor. As you explained, with phantom-refs enabled, if rt.jar is not available then it can appear that java.lang.Object is a subclass of both Serializable and Cloneable. This was leading to a loop where LCA(Array, Object) = Serializable and then elsewhere LCA(Serializable, Something Else) = Object.

I have tested a fix that replaces lines 148-151 of BytecodeHierarchy.java:

if ( ancestor_(RefType.v("java.io.Serializable"), rt) )
    r.add(RefType.v("java.io.Serializable"));
if ( ancestor_(RefType.v("java.lang.Cloneable"), rt) )
    r.add(RefType.v("java.lang.Cloneable"));

with:

/* Do not consider Object to be a subtype of Serializable or Cloneable
(it can appear this way if phantom-refs is enabled and rt.jar is not
available) otherwise an infinite loop can result. */
if (!TypeResolver.typesEqual(RefType.v("java.lang.Object"), rt)) {
    if ( ancestor_(RefType.v("java.io.Serializable"), rt) )
        r.add(RefType.v("java.io.Serializable"));
    if ( ancestor_(RefType.v("java.lang.Cloneable"), rt) )
        r.add(RefType.v("java.lang.Cloneable"));
}

This ensures that Serializable and Cloneable are only chosen as LCAs of an array type and a reference type if the reference type is not Object. This appears to resolve this particular issue, but I'm afraid I haven't had time to thoroughly check the behaviour of the type assigner with phantom-refs in all other respects.