See [Soot-list] patch to fix bug in SSA Phi Node Placement
"This is a patch to fix a bug in SSA Phi Node Placement. Running the
attached test case with no external Transformers causes a validation
error by the Java runtime. The patch makes node placement more in the
spirit of Cytron's original paper [1]. To eliminate extra phi nodes,
the patch does not place phi nodes is a local is only defined in one
block. The produced Jimple is comparatively as clean as before, but
without the bug."