Bugfix:
When -oaat and -outjar are used together, the output jar file remains empty.

That is because, even with -oaat flag, PackManager:writeOutput is called from soot.Main at line 199. I dont think there is no need for that call because with -oaat flag, the classes have already been output at that point.

The fix is to move that call to writeOutput from Soot.Main to inside the else branch in PackManager.runPacks() method.

Saswat