fixing bug in coffi.ClassFile with DataInputStream.read
On a mac DataInputStream.read is only reading part
of the array. The code is expecting the whole array
to be read so that the cursor is positioned properly
for the next read.

This is triggered when using -include-all on mac
with classes.jar and ui.jar and doing a body
transformer on:
  apple.awt.CDragSourceContextPeer
  com.apple.java.BackwardsCompatibilityTest