Fix for #120
Problem:
When a jar file contains both .class and .java files, the .java files
are picked up and given to the class loading algorithm to read. However,
the Java front-end cannot read from an archive and throws an exception.

The exception states that Polyglot isn't able to load the file, but
this can be confusing when the user is running Soot with Polyglot disabled.

Solution:
Ignore .java files in archives

Changes:
 * Ignored .java files in archives in SourceLocator
 * Improved the exception message in JavaClassProvider