Changed DominatorTree to have getHeads
DominatorTree's used to have getHead() as the method to get the head of
the tree. Unfortunately, if it was a post-dominator tree and you wanted
to compute the post-dominator frontier this caused problems for
functions with multiple distiguished exits (eg. more than one head in
the post-dom tree). Basically, what would happen is the frontier would
miss elements for certain graph constructions.

I have fixed this problem by removing the getHead() method and replacing
it with a getHeads() method. I also cleaned up a lot of whitespace
issues in these files.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>