Tests now build with Ant as well as Buildr
I was using buildr to test (as that is the build system I use to build
soot for my project). These updates allow the tests to be run from the
ant buildfile. Note, you need to add a Hamcrest jar and update your
junit.jar to something reasonably recent in the 4.x series (I recommend
just using junit-4.8.2.jar).

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>