Add check-init-throw-analysis option to Soot.
This option allows the user to specify a throw analysis to be used
for the checkInit() phase of code validation inside soot.Body.

Using '-check-init-throw-analysis dalvik' is a suitable fix for
this issue: https://github.com/Sable/soot/issues/109

Note that this adds a dependency on soot.dexpler to soot.Body.