Some (obfuscated?) APK files contain VirtualnvokeExprs to interface methods where they ought to use InterfaceInvokeExprs instead. This seems to be fine with dexopt, but fails during conversion to jimple.
To fix this, we add a check to the InvokeVirtualInstruction class such that it relays the jimplify() call to the InvokeInterfaceInstruction if the target is an interface. This is a bit of a hack, but the only alternative would be to have the factory perform the check, but the factory would then need a copy of the method reference parsing code which would have been even worse.