Making SmartLocalDefs even smarter by specially handling exceptional data flows. Assume that a statement a = foo() throws an exception. This means that the assignment is never carried out, so new value of a assigned at a = foo() will *not* be available inside the exception handler.
This also introduces a new variant of the intraprocedural analysis framework that not only allows to define flows over a statement, but to distinguish flow functions depending on the successor of the current statement.

The new variant of the analysis framework might need a bit more testing though it looks ok to me with the tests I have done so far.