Problem with the SmartLocalDefsPool: It recognizes if the units inside a body are changed and updated the cached localdefs analysis accordingly. It however does not notice if the contents of a box inside a unit in a body gets changed. However, since SmartLocalDefs only computes the local defs for variables v at units u if v is alive at u, a change of a use box after u might make v alive which it was not before. Thus, if we don't update the cached analysis results, we're missing a definition.
The problem is very generic, the solution is a hack for the CopyPropagator to manually invalidate the cache. This is a horrible hack since this would mean that every transformer which changesa box would need to know about the pool and invalidate it after it is done with its transformations. We definitely need a better solution here, but I'm trying to get things back to work for now, even if it's a bit cheesy.