Fixed an issue with catch-all traps and monitors. Whenever a monitor is held at a certain position in the code and this position may throw an exception, one must make sure that there is a trap which exits the monitor before leaving the method.
Some bytecode verifiers such as the one implemented in Android ART over-approximate this requirement and expect a catch-all clause around all statements that may throw exception and hold monitors. If Soot however detected that there was a trap for java.lang.Exception and one for java.lang.Throwable, it could remove the one for java.lang.Throwable if all possible exceptions are already caught by the one for java.lang.Exception. This then however failed the bytecode verifier which expected the catch-all in any case.

This fixes #339