Adds type-based resolution of invoke() calls
Adds an unsound approach for statically resolving calls to invoke().
The OnFlyCallGraphBuilder hooks into Spark and uses the reaching types
of the 'obj' argument and contents of the 'args' array to find methods
to invoke. In some cases it is possible to compute the positions in
the argument array of these reaching types, in which case the
resolution can be made more precise. This types-for-invoke approach
is still unsound:

* It doesn't handle arrays receivers
* It doesn't handle static methods
* It fails to account for default nulls in the argument array