Fixes dynamic class loading
When looking for loadable dynamic class, the File.pathSeparator
was used, which is the ":" or ";" character. As dynamic classes
are intended to be loaded from directories, this was almost
certainly meant to be "/" or "\". Also, if you had a non-directory
path on the soot classpath the dynamic class resolution would fail,
as Soot would attempt to load from the
directory /foo/bar/baz.jar/com/example/package.