Fixes type masks for classes with phantom parents
When a class has a parent that is a phantom class, it will not be
processed during the breadth first search in TypeManager, as it is technically
not reachable from the java.lang.Object type. This causes a missing type mask
that will lead cause an error for some use patterns: specifically if a class is
instantiated reflectively (via Soot's dynamic classes).

This commit also uses the built-in makeClasTypeMask method to traverse the
class hierarchy rooted at the
phantom-parent class.