Improve the time it takes to load classes from archives into soot.
This commit makes the following changes:
  1. Cache the entries in an archive on the class path so determining if
a class is in an archive requires only the IO operations needed to open the
archive and list all entries. Previously, determining if an entry was in the
archive required one to open the archive and query it each time.
  2. Cache the ClassSourceType of all entities on the class path.
When discovering the ClassSourceType, a check is performed to determine
if a path exists, is readable, is a file or directory, and if it was
a file what type. Previously, this check was performed each time an
entity on the class path was checked to determine if it contained a
class which was extremly redundent.
  3. Modify FoundFile to fit with the new caching scheme. This has the
added benifit of it not holding open an ZipFile until Soot actually
needs to read from the ZipFile.