Bug Fix: Setup the inner/outer class hierarchy when reading classes from jimple
Previously, the jimple parser completly omitted the setup of the inner/outer
class hierarchy as it read in classes from jimple. This just adds a bit of code
to the JimpleClassSource file that uses the class hierarchy laid out in a classes
full name to determine if it has an outer class, what the outer class is, and
to set the classes outer class. The name parsing procedure is the same as the
parsing procedure used in dexpler (i.e. it handles the jack and jill generated
lambda classes).