Added support for calls to Class.forName() /newInstance() to SPARK. When we have the target class name as a constant string, we create a class constant node in the PAG for it at the Class.forName() call. When this class constant node later appears in a call to Class.newInstance(), we create a node if the new type NewInstanceNode for it. This new node models that this is a dynamic alloc site: When constructing the PAG, we don't know the type yet, because it depends on the incoming class constants.
When we later do the propagation, we special-case the new node type. For all incoming class constant nodes that are propagated over a NewInstance node, we create one alloc node, because during propagation, we have all the type information we need.

This new feature is only enabled if types-for-invoke is enabled, so the default behavior of SPARK should not change.