Options and method source dropping
Added support for Boolean options that are true efault. In such a case,
we then allow for a command-line option no-<option name> to disable what
is enabled by default.

Also added an option to drop the method source once the body of that
method has been loaded once to save memory. This option is enabled by
default.