Remove nops before building Shimple or else Jimple validation may fail
Must remove nops prior to building the CFG because NopStmt appearing
before the IdentityStmt in a trap handler that is itself protected by a
trap cause Phi nodes to be inserted before the NopStmt and therefore
before the IdentityStmt. This introduces a validation problem if the Phi
nodes leave residual assignment statements after their removal.