Fix typing of long bit shift expressions in Grimp
The type of a bit shift expression only depends on its left operand -
the type of the right operand is always int. However, unlike the Jimple
variants (JShlExpr, JShrExpr and JUshrExpr), the Grimp shift expression
classes (GShlExpr, GShrExpr and GUshrExpr) do not override getType() and
use the default AbstractIntLongBinopExpr implementation, which sets the
type to long only if both operands are longs.

This causes a RuntimeException when converting long shifts from Java
bytecode to Grimp and then back to Java bytecode again, as the Grimp to
Java bytecode conversion code does not expect getType() to return
UnknownType.

This commit fixes the problem by copying the getType() method from the
Jimple shift expression classes to the equivalent Grimp classes.