Wrapped multi dex logic into MultiDexBuilder class. Readded interns that where removed when porting from DexBuilder to DexPool.
At this point I am not sure if it makes sense, to even make using multi dex optional since we will crash or otherwise produce an invalid apk anyway.

It might make sense to abstract from MultiDexBuilder to a custom DexBuilder class. We could use the abstract version in the DexPrinter and helper classes then and instanciate the Dexbuilder we are interested in in the constructor of DexPrinter depending on some options (e.g., SingleDexbuilder, MultiDexBuilder, SophisticatedMultiDexBuilder (for Android pre lolipop)). Other optimizations are possible as well, like packing classes in dex files euqaly named to those the classes where originally in.