Use global soot options when creating an ExceptionalUnitGraph
All JimpleBasedICFG were forced to use the UnitThrowAnalysis and and omit
exceptional unit edges when constructing an ExceptionalUnitGraph. There
are global settings for these options, so it makes more sense to use the
global settings than to have them hardcoded. Esepeically since this
exception to the global settings is not stated anywhere.