Supports AnySubType for reaching types
Our previous implementation only worked on concrete reference types.
Under Spark this was a MOSTLY a fine assumption, except that the native
simulators handlers could still introduce anysubtype types into the PAG.

This commit correctly handles this and, in principle, opens the
door for types-for-invoke to be used with the CHA builder.
However, the machinery to plug into the CHA is not written (yet).