Fixes types for init
The OFCGB is supposed to "subscribe" to updates of array contents
when that array flows to an invoke call. However, we were creating
the AllocAndDotField class by calling dot() on the alloc node,
which returns null if the field hasn't been explicitly alloced through
the PAG. This fixes the subscription mechanism to use pag's
makeAllocDotField.