fix NPE in soot.dexpler.DexBody.jimplify(Body b, SootMethod m)
Detailed:
If the last instruction of body is InvodeStaticInstruction, the last
dangling instruction's finalize will not be called. This leads to a
NullPointerException in the following fixLineNumbers() method. This
situation is not very common but still exists, e.g. some invokes
android.util.Log.i() at the end of the method body.