Add support for the invoke-polymorphic and invoke-custom instructions introduced in dex 38 (Android 8).
To mimic the invoke-dynamic instruction dex 38 introduced two different dynamic instructions
invoke-polymorphic and invoke-custom. Where as invoke-custom behaves similar to if not exactly
like invoke-dynamic, invoke-polymorphic is a weaker version.

More information on these two instructions can be found here.

https://www.pnfsoftware.com/blog/android-o-and-dex-version-38-new-dalvik-opcodes-to-support-dynamic-invocation/

These changes handle translating these two instructions in dex back into something equivalent
in normal bytecode. Note, the most recent build of dexlib2 does not properly support either of
these instructions. The framework is there so you will not have any compilation problems.
However, there are a number of bugs and implitation issues because their code was based off of
a beta version of the instructions. I have submited fixes to smali git repo and they should be
pulled in the repo soon. However, I do not know when he plans on pushing a new version to maven.

Examples of proper invoke-polymorphic and invoke-custom instructions can be found in the tests
used for dexdump in the aosp repo (i.e. invoke-custom.dex and invoke-polymorphic.dex).

https://android.googlesource.com/platform/art/+/android-8.1.0_r41/test/dexdump/