Implement parameter adaptation in LambdaMetaFactory
Implemented modeling of parameter adaptation. This is required to
handle conversion of primitives to their wrapper types
(boxing and unboxing).

Enabled writing generated classes to another format, such as .class.

Generated class names follow an inner class naming pattern.

Removed BridgeMethodSource; ThunkMethodSource handles all cases.
The samMethodType and the bridges are
signatures to be implemented by the functional object,
and each one delegates to the implMethod.