Make the UnconditionalBranchFolder class stateless to avoid concurrency issues
Since commit 310e6223ed751cba9f8b8a545a03e524918f5531 ("even more parallel
processing"), the classes in PackManager.runBodyPacks(...) are potentially
processed by multiple threads. Consequently, the transformers should be
thread-safe.
The UnconditionalBranchFolder class, which is part of the "jop" pack, is not
thread-safe, which can result in a NPE:

Excerpt:
java.lang.NullPointerException
        at soot.jimple.toolkits.scalar.UnconditionalBranchFolder.internalTransform(UnconditionalBranchFolder.java:74)
        at soot.BodyTransformer.transform(BodyTransformer.java:51)
        at soot.Transform.apply(Transform.java:120)
        at soot.BodyPack.internalApply(BodyPack.java:50)
        at soot.Pack.apply(Pack.java:125)
        at soot.PackManager.runBodyPacks(PackManager.java:1009)
        at soot.PackManager.lambda$runBodyPacks$0(PackManager.java:660)
        at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1128)
        at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:628)
        at java.base/java.lang.Thread.run(Thread.java:834)

Instead of storing the "state" in instance variables, the state is stored
in local variables and passed to other methods via parameters methods as
needed.

Signed-off-by: Ingo Budde <ingo.budde@iem.fraunhofer.de>
Signed-off-by: Marcus Hüwe <marcus.huewe@iem.fraunhofer.de>