Fix potential race in Scence.getSootClassUnsafe
There is a potential race condition in Scene.getSootClassUnsafe

java.lang.RuntimeException: Failed to convert <Foo: void publish(byte[])>
    at soot.asm.AsmMethodSource.getBody(AsmMethodSource.java:2163)
    at soot.SootMethod.retrieveActiveBody(SootMethod.java:402)
    at soot.PackManager.runBodyPacks(PackManager.java:998)
    at soot.PackManager.lambda$runBodyPacks$0(PackManager.java:660)
    at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1128)
    at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:628)
    at java.base/java.lang.Thread.run(Thread.java:834)
Caused by: java.lang.RuntimeException: duplicate class: java.lang.invoke.StringConcatFactory
    at soot.Scene.addClassSilent(Scene.java:892)
    at soot.Scene.getSootClassUnsafe(Scene.java:1230)
    at soot.Scene.getSootClassUnsafe(Scene.java:1202)
    at soot.asm.AsmMethodSource.getClassFromScene(AsmMethodSource.java:358)
    at soot.asm.AsmMethodSource.toSootMethodRef(AsmMethodSource.java:1571)
    at soot.asm.AsmMethodSource.convertInvokeDynamicInsn(AsmMethodSource.java:1459)
    at soot.asm.AsmMethodSource.convert(AsmMethodSource.java:1909)
    at soot.asm.AsmMethodSource.getBody(AsmMethodSource.java:2161)
    ... 6 more

For this, consider the following two threads:

Thread 1                                        Thread 2

  type = getOrAddRefType(className);
  synchronized (type) {
    SootClass c = new SootClass(className);
    c.isPhantom = true;
    addClassSilent(c);
    ...
  }

                                                type = getOrAddRefType(className);
                                                synchronized (type) {
                                                  SootClass c = new SootClass(className);
                                                  c.isPhantom = true;
               the class already exists --->      addClassSilent(c);

To fix this, check if the type already has a SootClass (this has
to happen under the lock/monitor).
It is important to note that the type.getSootClass() call must not
be executed under the "type" lock in order to avoid a potential
deadlock (Scene.getSootClass(x, y) -> RefType.getSootClass ->
SootResolver.makeClassRef -> Scene.getSootClassUnsafe(x) ->
Scene.getSootClassUnsafe(x, y)) (the deadlock cannot occur in the
current Scene class but we have know clue if a potential
subclass removes an element from the nameToClass map).

Note: Even though we avoid a potential deadlock, we do not avoid a
potential "inifite" recursion. That's OK because the old code already
had the potential "infinite" recursion.

Note 2: We have not checked if there are additional potential races
if the ModuleUtil.module_mode() is enabled.

Signed-off-by: Marcus Hüwe <marcus.huewe@iem.fraunhofer.de>