Add DirectedBodyGraph interface
Several of the DirectedGraph implementations involve a method Body and provide a getBody() method to obtain a reference to it. It is often useful to treat these graphs in the same manner and be able to obtain that Body reference. However, the DirectedGraph interface does not (and should not) define that method. Thus, the DirectedBodyGraph interface is introduced to the hierarchy of these classes that provide the getBody() method.