Cleanup shimple classes, and fix a couple bugs
In several classes, removed the fields caching objects from the ShimpleFactory because in a scenario where the body was modified and the factory was reset, these fields would contain stale data (and there was little benefit to caching them anyway). Also removed unused fiels from DefaultShimpleFactory and the unused GuaranteedDefs from PhiNodeManager.