bugfix: make explodeClassPath(String) treat empty paths consistently
Empty paths (i.e. substrings separated by File.pathSeparator) appearing at the end of the parameter String would be ignored (due to the behavior of String.split(..)) but those that were not at the end would be included resulting in the current working directory being added to the exploded classpath.