When preserving original names, use the local variable table for guidance
When the bytecode LocalVariableTable associates different names with the same index at different points in the program (or the same name to different locals), the SharedInitializationLocalSplitter (maybe even the normal LocalSplitter) can end up with different versions of the same local name that cannot be combined by the LocalPacker and thus will output bytecode using the '#' character in the name of locals and that is disallowed by some JRE implementations. This commit fixes the issue by assigning different names to all of the locals from the start when LocalVariableTable information is still available and thus each unique local name can be as close to the original local as possible.