added support for passing on types in the PAG in case we later want to later invoke a method that object which we have a summary.
Example:

List<...> l = ...;
Spliterator<...> s = l.spliterator();
s.forEachRemaining(callback);

We would normally lose the type information over the call to spliterator() if we don't have an implementation for that method. With the new deferred PAG summary, we can just use the declared type because we know that we will use a summary for the call to forEachRemaining() later anyway, so the actual type of the Spliterator is irrelevant.